+++ src/lib/Components/Bidding/__stories__/BidFlow.story.tsx
@@ -6,15 +6,8 @@ import { BidFlowRenderer } from "lib/relay/QueryRenderers"
 import renderWithLoadProgress from "lib/utils/renderWithLoadProgress"
 import createEnvironment from "../../../relay/createEnvironment"
 
-import { Flex } from "../Elements/Flex"
-import { Sans12, Serif14, Serif16 } from "../Elements/Typography"
-
-import { BiddingThemeProvider } from "../Components/BiddingThemeProvider"
-import { Checkbox } from "../Components/Checkbox"
-
 import { NavigatorIOS } from "react-native"
 import BidFlow from "../../../Containers/BidFlow"
-import { Input } from "../Components/Input"
 import { BidResultScreen } from "../Screens/BidResult"
 import { BillingAddress } from "../Screens/BillingAddress"
 import { ConfirmBid } from "../Screens/ConfirmBid"
@@ -48,13 +41,21 @@ const BidFlowStoryRenderer: React.SFC<any> = ({ render, query, saleArtworkID }) 
 
 storiesOf("Bidding")
   .add("Show bid flow", () => {
-    return <BidFlowRenderer render={renderWithLoadProgress(BidFlow)} artworkID={testArtworkID} saleID={testSaleID} />
+    return (
+      <BidFlowRenderer
+        render={renderWithLoadProgress(BidFlow)}
+        artworkID={testArtworkID}
+        saleID={testSaleID}
+        intent="bid"
+      />
+    )
   })
   .add("Select Max Bid", () => (
     <BidFlowStoryRenderer
       render={renderWithLoadProgress(MaxBidScreen)}
       query={selectMaxBidQuery}
       saleArtworkID={testSaleArtworkID}
+      intent="bid"
     />
   ))
   .add("Confirm Bid", () => {
@@ -63,6 +64,8 @@ storiesOf("Bidding")
         sale_artwork={{
           sale: {
             id: "1",
+            live_start_at: "2018-06-11T01:00:00+00:00",
+            end_at: null,
           },
           artwork: {
             id: "1",
@@ -87,6 +90,8 @@ storiesOf("Bidding")
             sale_artwork: {
               sale: {
                 id: "1",
+                live_start_at: "2018-06-11T01:00:00+00:00",
+                end_at: null,
               },
               artwork: {
                 id: "1",
@@ -113,14 +118,18 @@ storiesOf("Bidding")
         })}
         query={bidResultQuery}
         saleArtworkID={testSaleArtworkID}
+        intent="bid"
       />
     )
   })
   .add("Bidding Result (not highest bid)", () => {
     const status = "OUTBID"
     const messageHeader = "Your bid wasn’t high enough"
-    const messageDescriptionMd = `Another bidder placed a higher max bid or the same max bid before you did.  \
- Bid again to take the lead.`
+    const messageDescriptionMd = `
+      Another bidder placed a higher max bid or the same max bid before you did.
+      Bid again to take the lead.
+    `
+
     return (
       <BidFlowStoryRenderer
         render={renderWithLoadProgress(BidResultScreen, {
@@ -131,11 +140,12 @@ storiesOf("Bidding")
         })}
         query={bidResultQuery}
         saleArtworkID={testSaleArtworkID}
+        intent="bid"
       />
     )
   })
   .add("Credit Card", () => {
-    return <CreditCardForm />
+    return <CreditCardForm onSubmit={params => console.warn(params)} />
   })
   .add("Billing Address", () => {
     return <BillingAddress />
@@ -143,8 +153,11 @@ storiesOf("Bidding")
   .add("Bidding Result (live bidding started)", () => {
     const status = "ERROR_LIVE_BIDDING_STARTED"
     const messageHeader = "Live bidding has started"
-    const messageDescriptionMd = `Sorry, your bid wasn’t received before live bidding started.\
- To continue bidding, please [join the live auction](http://live-staging.artsy.net/).`
+    const messageDescriptionMd = `
+      Sorry, your bid wasn’t received before live bidding started.
+      To continue bidding, please [join the live auction](http://live-staging.artsy.net/).
+    `
+
     return (
       <BidFlowStoryRenderer
         render={renderWithLoadProgress(BidResultScreen, {
@@ -155,61 +168,7 @@ storiesOf("Bidding")
         })}
         query={bidResultQuery}
         saleArtworkID={testSaleArtworkID}
+        intent="bid"
       />
     )
   })
-
-storiesOf("App Style/Input")
-  .add("Text Input", () => (
-    <BiddingThemeProvider>
-      <Flex mt={7} ml={4} mr={4}>
-        <Serif16 mb={2}>Title</Serif16>
-        <Input placeholder="Placeholder" mb={5} />
-
-        <Serif16>Title</Serif16>
-        <Serif14 mb={2} color="black60">
-          Short description
-        </Serif14>
-        <Input placeholder="Placeholder" value="Content" mb={5} />
-
-        <Input placeholder="Without Title" mb={5} />
-
-        <Serif16 mb={2}>Error</Serif16>
-        <Input error placeholder="Placeholder" mb={3} />
-        <Sans12 color="red100">Error message</Sans12>
-      </Flex>
-    </BiddingThemeProvider>
-  ))
-  .add("Check Boxes", () => (
-    <BiddingThemeProvider>
-      <Flex mt={7}>
-        <Checkbox pl={3} pb={1}>
-          <Serif16 mt={2}>Remember me</Serif16>
-        </Checkbox>
-
-        <Checkbox pl={3} pb={1} checked>
-          <Serif16 mt={2}>Remember me</Serif16>
-        </Checkbox>
-
-        <Checkbox pl={3} pb={1} error>
-          <Serif16 mt={2} color="red100">
-            Agree to Terms and Conditions
-          </Serif16>
-        </Checkbox>
-
-        <Checkbox pl={3} pb={1} checked error>
-          <Serif16 mt={2} color="red100">
-            Agree to Terms and Conditions
-          </Serif16>
-        </Checkbox>
-
-        <Checkbox pl={3} pb={1} disabled>
-          <Serif16 mt={2}>Remember me</Serif16>
-        </Checkbox>
-
-        <Checkbox pl={3} pb={1} checked disabled>
-          <Serif16 mt={2}>Remember me</Serif16>
-        </Checkbox>
-      </Flex>
-    </BiddingThemeProvider>
-  ))